backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pydantic-settings
sqlalchemy[asyncio]
asyncpg
aiosqlite
alembic
PyJWT
bcrypt
python-multipart
python-dotenv
httpx
spacy
numpy
pandas
weasyprint
supabase
slowapi
pytest
pytest-asyncio
pytest-cov
anyio

